Rename directory `app/` to `application/`
