+++ application/action.go
@@ -1,4 +1,4 @@
-package app
+package application
 
 func (a *ExitAction) Perform() error {
 	return nil
