+++ application/command_test.go
@@ -1,4 +1,4 @@
-package app
+package application
 
 import (
 	"errors"
