+++ application/types.go
@@ -1,4 +1,4 @@
-package app
+package application
 
 import "github.com/Rindrics/require-label-prefix/domain"
