+++ infra/client.go
@@ -3,7 +3,7 @@ package infra
 import (
 	"context"
 
-	"github.com/Rindrics/require-label-prefix/app"
+	"github.com/Rindrics/require-label-prefix/application"
 	"github.com/google/go-github/github"
 	"golang.org/x/oauth2"
 )
@@ -21,14 +21,15 @@ func NewGitHubClient(token string) *GitHubClient {
 	return &GitHubClient{client: github.NewClient(tc)}
 }
 
-func (g *GitHubClient) PostComment(p app.PostCommentParams) error {
+func (g *GitHubClient) PostComment(p application.PostCommentParams) error {
 	body := "comment body"
 	comment := &github.IssueComment{Body: &body}
 	_, _, err := g.client.Issues.CreateComment(context.Background(), p.RepoInfo.Owner, p.RepoInfo.Repo, p.Number, comment)
 	return err
 }
 
-func (g *GitHubClient) AddLabels(p app.AddLabelsParams) error {
+func (g *GitHubClient) AddLabels(p application.AddLabelsParams) error {
+
 	_, _, err := g.client.Issues.AddLabelsToIssue(context.Background(), p.RepoInfo.Owner, p.RepoInfo.Repo, p.Number, p.Labels)
 
 	return err
